Core: Adding Performance, Utils\TickCount is now handled by Stopwatch->Timestamp, Logging function name attirbute added.

+++ Core/Utils/DelayAction.cs
@@ -78,7 +78,7 @@ namespace LeagueSharp.CommonEx.Core.Utils
             /// <summary>
             ///     Time.
             /// </summary>
-            public int Time;
+            public long Time;
 
             /// <summary>
             ///     Action Constructor.
